# perf(encoding): use estimate size to reserve memory in row encoding